backend/requirements.txt
flask
flask-restful
flask-cors
gunicorn
flask-jwt-extended
flask-sqlalchemy
psycopg2-binary
requests
google-cloud-storage
python-docx
pandas
openpyxl
opencv-python-headless